Move query-editor
